requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.6.4
